+++ cooler/cli/__init__.py
@@ -2,18 +2,46 @@
 from __future__ import division, print_function
 import logging
 import sys
-
-import click
 from .. import __version__, get_logger
-
-logging.basicConfig(stream=sys.stderr)
-logging.captureWarnings(True)
-logger = get_logger()
-logger.setLevel(logging.INFO)
+import click
 
 
 # Monkey patch
 click.core._verify_python3_env = lambda: None
+
+
+class UnsortedGroup(click.Group):
+    def list_commands(self, ctx):
+        return list(self.commands)
+
+    # def format_commands(self, ctx, formatter):
+    #     """Extra format methods for multi methods that adds all the commands
+    #     after the options.
+    #     """
+    #     commands = []
+    #     for subcommand in self.list_commands(ctx):
+    #         cmd = self.get_command(ctx, subcommand)
+    #         # What is this, the tool lied about a command.  Ignore it
+    #         if cmd is None:
+    #             continue
+    #         if cmd.hidden:
+    #             continue
+
+    #         commands.append((subcommand, cmd))
+
+    #     # allow for 3 times the default spacing
+    #     if len(commands):
+    #         limit = formatter.width - 6 - max(len(cmd[0]) for cmd in commands)
+
+    #         rows = []
+    #         for subcommand, cmd in commands:
+    #             help = cmd.get_short_help_str(limit)
+    #             rows.append((subcommand, help))
+
+    #         if rows:
+    #             with formatter.section('Commands'):
+    #                 formatter.write_dl(rows)
+
 
 
 CONTEXT_SETTINGS = {
@@ -22,14 +50,14 @@ CONTEXT_SETTINGS = {
 
 
 @click.version_option(version=__version__)
-@click.group(context_settings=CONTEXT_SETTINGS)
+@click.group(context_settings=CONTEXT_SETTINGS, cls=UnsortedGroup)
 @click.option(
-    '--debug/--no-debug', 
-    help="Verbose logging", 
+    '--debug/--no-debug',
+    help="Verbose logging",
     default=False)
 @click.option(
-    '-pm', '--post-mortem', 
-    help="Post mortem debugging", 
+    '-pm', '--post-mortem',
+    help="Post mortem debugging",
     is_flag=True,
     default=False)
 def cli(debug, post_mortem):
@@ -37,8 +65,13 @@ def cli(debug, post_mortem):
     Type -h or --help after any subcommand for more information.
 
     """
+    logging.basicConfig(stream=sys.stderr)
+    logging.captureWarnings(True)
+    root_logger = get_logger()
     if debug:
-        logger.setLevel(logging.DEBUG)
+        root_logger.setLevel(logging.DEBUG)
+    else:
+        root_logger.setLevel(logging.INFO)
 
     if post_mortem:
         import traceback
@@ -54,17 +87,17 @@ def cli(debug, post_mortem):
 
 
 from . import (
+    cload,
+    load,
+    info,
+    dump,
+    show,
+    balance,
+    merge,
+    coarsen,
+    zoomify,
     makebins,
     digest,
     csort,
-    cload,
-    load,
-    merge,
-    copy,
-    list_,
-    info,
-    dump,
-    balance,
-    aggregate,
-    show,
+    fileops
 )
