+++ cooler/io/ingest.py
@@ -3,7 +3,7 @@
 Contact Binners
 ~~~~~~~~~~~~~~~
 
-Binners are iterators that convert input data of various flavors into a 
+Binners are iterators that convert input data of various flavors into a
 properly sorted, chunked stream of binned contacts.
 
 """
@@ -22,9 +22,11 @@ import numpy as np
 import pandas as pd
 import h5py
 
-from .utils import GenomeSegmentation, parse_cooler_uri, balanced_partition
 from .. import get_logger
-from ..util import rlencode, get_binsize, get_chromsizes, parse_region
+from ..util import (
+    parse_region, rlencode, get_binsize, get_chromsizes, GenomeSegmentation,
+    balanced_partition
+)
 from ..tools import lock, partition
 
 
@@ -36,22 +38,22 @@ class BadInputError(ValueError):
 
 
 SANITIZE_PRESETS = {
-    'bg2': dict(decode_chroms=True, is_one_based=False,  tril_action='reflect', 
-                chrom_field='chrom', anchor_field='start', 
-                sided_fields=('chrom', 'start', 'end'), suffixes=('1', '2'), 
+    'bg2': dict(decode_chroms=True, is_one_based=False,  tril_action='reflect',
+                chrom_field='chrom', anchor_field='start',
+                sided_fields=('chrom', 'start', 'end'), suffixes=('1', '2'),
                 sort=True, validate=True),
 
-    'pairs': dict(decode_chroms=True, is_one_based=False, tril_action='reflect', 
-                  chrom_field='chrom', anchor_field='pos', 
+    'pairs': dict(decode_chroms=True, is_one_based=False, tril_action='reflect',
+                  chrom_field='chrom', anchor_field='pos',
                   sided_fields=('chrom', 'pos'), suffixes=('1', '2'),
                   sort=False, validate=True)
 }
 
 
 
-def _sanitize_records(chunk, gs, decode_chroms, is_one_based, tril_action, 
-                      chrom_field, anchor_field, sided_fields, suffixes, 
-                      sort, validate):    
+def _sanitize_records(chunk, gs, decode_chroms, is_one_based, tril_action,
+                      chrom_field, anchor_field, sided_fields, suffixes,
+                      sort, validate):
     # Get integer contig IDs
     if decode_chroms:
         # Unspecified chroms get assigned category = NaN and integer code = -1
@@ -63,12 +65,12 @@ def _sanitize_records(chunk, gs, decode_chroms, is_one_based, tril_action,
         chrom1_ids = chunk['chrom1'].values
         chrom2_ids = chunk['chrom2'].values
         if validate:
-            for col, dt in [('chrom1', chrom1_ids.dtype), 
+            for col, dt in [('chrom1', chrom1_ids.dtype),
                             ('chrom2', chrom2_ids.dtype)]:
                 if not is_integer_dtype(dt):
                     raise BadInputError(
-                        "`{}` column is non-integer. ".format(col) + 
-                        "If string, use `decode_chroms=True` to convert to enum")    
+                        "`{}` column is non-integer. ".format(col) +
+                        "If string, use `decode_chroms=True` to convert to enum")
 
     # Drop records from non-requested chromosomes
     to_drop = (chrom1_ids < 0) | (chrom2_ids < 0)
@@ -90,20 +92,20 @@ def _sanitize_records(chunk, gs, decode_chroms, is_one_based, tril_action,
     if is_one_based:
         anchor1 -= 1
         anchor2 -= 1
-    
+
     # Check types and bounds
     if validate:
         for dt in [anchor1.dtype, anchor2.dtype]:
             if not is_integer_dtype(dt):
                 raise BadInputError("Found a non-integer anchor column")
-        
+
         is_neg = (anchor1 < 0) | (anchor2 < 0)
         if np.any(is_neg):
             err = chunk[is_neg]
             raise BadInputError(
                 "Found an anchor position with negative value:\n{}".format(
                 err.head().to_csv(sep='\t')))
-        
+
         chromsizes1 = gs.chromsizes[chrom1_ids].values
         chromsizes2 = gs.chromsizes[chrom2_ids].values
         is_excess = (anchor1 > chromsizes1) | (anchor2 > chromsizes2)
@@ -116,7 +118,7 @@ def _sanitize_records(chunk, gs, decode_chroms, is_one_based, tril_action,
     # Handle lower triangle records
     if tril_action is not None:
         is_tril = (
-            (chrom1_ids > chrom2_ids) | 
+            (chrom1_ids > chrom2_ids) |
             ((chrom1_ids == chrom2_ids) & (anchor1 > anchor2))
         )
         if np.any(is_tril):
@@ -125,7 +127,7 @@ def _sanitize_records(chunk, gs, decode_chroms, is_one_based, tril_action,
                     chrom2_ids[is_tril], chrom1_ids[is_tril]
                 anchor1[is_tril], anchor2[is_tril] = \
                     anchor2[is_tril], anchor1[is_tril]
-                for field in sided_fields:                
+                for field in sided_fields:
                     chunk.loc[is_tril, field + suffixes[0]], \
                     chunk.loc[is_tril, field + suffixes[1]] = \
                         chunk.loc[is_tril, field + suffixes[1]], \
@@ -153,22 +155,22 @@ def _sanitize_records(chunk, gs, decode_chroms, is_one_based, tril_action,
         start_abspos = gs.start_abspos
         bin1_ids = []
         bin2_ids = []
-        for cid1, pos1, cid2, pos2 in zip(chrom1_ids, anchor1, 
+        for cid1, pos1, cid2, pos2 in zip(chrom1_ids, anchor1,
                                           chrom2_ids, anchor2):
             lo = chrom_binoffset[cid1]
             hi = chrom_binoffset[cid1 + 1]
             bin1_ids.append(lo + np.searchsorted(
-                                start_abspos[lo:hi], 
+                                start_abspos[lo:hi],
                                 chrom_abspos[cid1] + pos1,
                                 side='right') - 1)
             lo = chrom_binoffset[cid2]
             hi = chrom_binoffset[cid2 + 1]
             bin2_ids.append(lo + np.searchsorted(
-                                start_abspos[lo:hi], 
+                                start_abspos[lo:hi],
                                 chrom_abspos[cid2] + pos2,
                                 side='right') - 1)
         chunk['bin1_id'] = bin1_ids
-        chunk['bin2_id'] = bin2_ids  
+        chunk['bin2_id'] = bin2_ids
     else:
         chunk['bin1_id'] = chrom_binoffset[chrom1_ids] + anchor1 // binsize
         chunk['bin2_id'] = chrom_binoffset[chrom2_ids] + anchor2 // binsize
@@ -176,39 +178,39 @@ def _sanitize_records(chunk, gs, decode_chroms, is_one_based, tril_action,
     # Sort by bin IDs
     if sort:
         chunk = chunk.sort_values(['bin1_id', 'bin2_id'])
-    
+
     return chunk
 
 
 def sanitize_records(bins, schema=None, **kwargs):
     """
-    Generates a funtion to sanitize and assign bin IDs to a data frame of 
+    Generates a funtion to sanitize and assign bin IDs to a data frame of
     paired genomic positions based on a provided genomic bin segmentation.
-    
+
     Parameters
     ----------
     bins : DataFrame
         Bin table to compare records against.
     schema : str, optional
-        Use pre-defined parameters for a particular format. Any options can be 
-        overriden via kwargs. If not provided, values for all the options below 
+        Use pre-defined parameters for a particular format. Any options can be
+        overriden via kwargs. If not provided, values for all the options below
         must be given.
-    
+
     decode_chroms : bool
-        Convert string chromosome names to integer IDs based on the order given 
-        in the bin table. Set to False if the chromosomes are already given as 
-        an enumeration, starting at 0. Records with either chrom ID < 0 are 
+        Convert string chromosome names to integer IDs based on the order given
+        in the bin table. Set to False if the chromosomes are already given as
+        an enumeration, starting at 0. Records with either chrom ID < 0 are
         dropped.
     is_one_based : bool
-        Whether the input anchor coordinates are one-based, rather than 
+        Whether the input anchor coordinates are one-based, rather than
         zero-based. They will be converted to zero-based.
     tril_action : 'reflect', 'drop', 'raise' or None
         How to handle lower triangle ("tril") records.
         If set to 'reflect', tril records will be flipped or "reflected"
         to their mirror image: "sided" column pairs will have their values
         swapped.
-        If set to 'drop', tril records will be discarded. This is useful if 
-        your input data is symmetric, i.e. contains mirror duplicates of every 
+        If set to 'drop', tril records will be discarded. This is useful if
+        your input data is symmetric, i.e. contains mirror duplicates of every
         record.
         If set to 'raise', an exception will be raised if any tril record is
         encountered.
@@ -217,7 +219,7 @@ def sanitize_records(bins, schema=None, **kwargs):
     anchor_field : str
         Base name of the positional anchor columns.
     sided_fields : sequence of str
-        Base names of column pairs to swap values between when 
+        Base names of column pairs to swap values between when
         mirror-reflecting records.
     suffixes : pair of str
         Suffixes used to identify pairs of sided columns. e.g.: ('1', '2'),
@@ -227,12 +229,12 @@ def sanitize_records(bins, schema=None, **kwargs):
     validate : bool
         Whether to do type- and bounds-checking on the anchor position
         columns. Raises BadInputError.
-        
+
     Returns
     -------
-    Function of one argument that takes a raw dataframe and returns a sanitized 
+    Function of one argument that takes a raw dataframe and returns a sanitized
     dataframe with bin IDs assigned.
-    
+
     """
     if schema is not None:
         try:
@@ -247,9 +249,9 @@ def sanitize_records(bins, schema=None, **kwargs):
     return partial(_sanitize_records, **options)
 
 
-def _sanitize_pixels(chunk, gs, is_one_based=False, tril_action='reflect', 
-                    bin1_field='bin1_id', bin2_field='bin2_id', sided_fields=(), 
-                    suffixes=('1', '2'), validate=True, sort=True):        
+def _sanitize_pixels(chunk, gs, is_one_based=False, tril_action='reflect',
+                    bin1_field='bin1_id', bin2_field='bin2_id', sided_fields=(),
+                    suffixes=('1', '2'), validate=True, sort=True):
     if is_one_based:
         chunk['bin1_id'] -= 1
         chunk['bin2_id'] -= 1
@@ -261,7 +263,7 @@ def _sanitize_pixels(chunk, gs, is_one_based=False, tril_action='reflect',
                 chunk.loc[is_tril, 'bin1_id'], \
                     chunk.loc[is_tril, 'bin2_id'] = chunk.loc[is_tril, 'bin2_id'], \
                                                        chunk.loc[is_tril, 'bin1_id']
-                for field in sided_fields:                
+                for field in sided_fields:
                     chunk.loc[is_tril, field + suffixes[0]], \
                     chunk.loc[is_tril, field + suffixes[1]] = \
                         chunk.loc[is_tril, field + suffixes[1]], \
@@ -283,16 +285,16 @@ def _validate_pixels(chunk, n_bins, boundscheck, triucheck, dupcheck, ensure_sor
         if np.any(is_neg):
             raise BadInputError("Found bin ID < 0")
         is_excess = (chunk['bin1_id'] >= n_bins) | (chunk['bin2_id'] >= n_bins)
-        if np.any(is_excess): 
+        if np.any(is_excess):
             raise BadInputError(
-                "Found a bin ID that exceeds the declared number of bins. " 
+                "Found a bin ID that exceeds the declared number of bins. "
                 "Check whether your bin table is correct.")
-    
+
     if triucheck:
         is_tril = chunk['bin1_id'] > chunk['bin2_id']
         if np.any(is_tril):
             raise BadInputError("Found bin1_id greater than bin2_id")
-    
+
     if not isinstance(chunk, pd.DataFrame):
         chunk = pd.DataFrame(chunk)
 
@@ -304,16 +306,16 @@ def _validate_pixels(chunk, n_bins, boundscheck, triucheck, dupcheck, ensure_sor
 
     if ensure_sorted:
         chunk = chunk.sort_values(['bin1_id', 'bin2_id'])
-    
+
     return chunk
 
 
 def validate_pixels(n_bins, boundscheck, triucheck, dupcheck, ensure_sorted):
     return partial(
-        _validate_pixels, 
-        n_bins=n_bins, 
-        boundscheck=boundscheck, 
-        triucheck=triucheck, 
+        _validate_pixels,
+        n_bins=n_bins,
+        boundscheck=boundscheck,
+        triucheck=triucheck,
         dupcheck=dupcheck,
         ensure_sorted=ensure_sorted)
 
@@ -327,28 +329,28 @@ def sanitize_pixels(bins, **kwargs):
     ----------
     bins : DataFrame
         Bin table to compare pixel records against.
-    
+
     is_one_based : bool, optional
-        Whether the input bin IDs are one-based, rather than zero-based. 
+        Whether the input bin IDs are one-based, rather than zero-based.
         They will be converted to zero-based.
     tril_action : 'reflect', 'drop', 'raise' or None
         How to handle lower triangle ("tril") pixels.
-        If set to 'reflect' [default], tril pixels will be flipped or 
-        "reflected" to their mirror image: "sided" column pairs will have their 
+        If set to 'reflect' [default], tril pixels will be flipped or
+        "reflected" to their mirror image: "sided" column pairs will have their
         values swapped.
-        If set to 'drop', tril pixels will be discarded. This is useful if 
-        your input data is symmetric, i.e. contains mirror duplicates of every 
+        If set to 'drop', tril pixels will be discarded. This is useful if
+        your input data is symmetric, i.e. contains mirror duplicates of every
         record.
         If set to 'raise', an exception will be raised if any tril record is
         encountered.
     bin1_field : str
-        Name of the column representing ith (row) axis of the matrix. 
+        Name of the column representing ith (row) axis of the matrix.
         Default is 'bin1_id'.
     bin2_field : str
-        Name of the column representing jth (col) axis of the matrix. 
+        Name of the column representing jth (col) axis of the matrix.
         Default is 'bin2_id'.
     sided_fields : sequence of str
-        Base names of column pairs to swap values between when mirror-reflecting 
+        Base names of column pairs to swap values between when mirror-reflecting
         pixels.
     suffixes : pair of str
         Suffixes used to identify pairs of sided columns. e.g.: ('1', '2'),
@@ -356,12 +358,12 @@ def sanitize_pixels(bins, **kwargs):
     sort : bool
         Whether to sort the output dataframe by bin_id and bin2_id.
     validate : bool
-        Whether to do type- and bounds-checking on the bin IDs. 
+        Whether to do type- and bounds-checking on the bin IDs.
         Raises BadInputError.
-        
+
     Returns
     -------
-    Function of one argument that takes a raw dataframe and returns a sanitized 
+    Function of one argument that takes a raw dataframe and returns a sanitized
     dataframe.
 
     """
@@ -527,16 +529,16 @@ class TabixAggregator(ContactBinner):
         import dill
         import pickle
         dill.settings['protocol'] = pickle.HIGHEST_PROTOCOL
-        
+
         self._map = map
         self.n_chunks = n_chunks
         self.is_one_based = bool(is_one_based)
         self.C2 = kwargs.pop('C2', 3)
         self.P2 = kwargs.pop('P2', 4)
-        
+
         # all requested contigs will be placed in the output matrix
         self.gs = GenomeSegmentation(chromsizes, bins)
-        
+
         # find available contigs in the contact list
         self.filepath = filepath
         self.n_records = None
@@ -547,14 +549,14 @@ class TabixAggregator(ContactBinner):
                 self.file_contigs = f.contigs
             if not len(self.file_contigs):
                 raise RuntimeError("No reference sequences found.")
-        
+
         # warn about requested contigs not seen in the contact list
         for chrom in self.gs.contigs:
             if chrom not in self.file_contigs:
                 warnings.warn(
                     "Did not find contig " +
                     " '{}' in contact list file.".format(chrom))
-        
+
         warnings.warn(
             "NOTE: When using the Tabix aggregator, make sure the order of "
             "chromosomes in the provided chromsizes agrees with the chromosome "
@@ -575,35 +577,35 @@ class TabixAggregator(ContactBinner):
         P2 = self.P2
 
         logger.info('Binning {}:{}-{}|*'.format(chrom1, start, end))
-        
+
         these_bins = self.gs.fetch((chrom1, start, end))
         rows = []
         with pysam.TabixFile(filepath, 'r', encoding='ascii') as f:
             parser = pysam.asTuple()
             accumulator = Counter()
-            
+
             for bin1_id, bin1 in these_bins.iterrows():
                 for line in f.fetch(chrom1, bin1.start, bin1.end,
                                     parser=parser):
                     chrom2 = line[C2]
                     pos2 = int(line[P2]) - decr
-                    
+
                     try:
                         cid2 = idmap[chrom2]
                     except KeyError:
                         # this chrom2 is not requested
                         continue
-                    
+
                     if binsize is None:
                         lo = chrom_binoffset[cid2]
                         hi = chrom_binoffset[cid2 + 1]
                         bin2_id = lo + np.searchsorted(
-                            start_abspos[lo:hi], 
+                            start_abspos[lo:hi],
                             chrom_abspos[cid2] + pos2,
                             side='right') - 1
                     else:
                         bin2_id = chrom_binoffset[cid2] + (pos2 // binsize)
-                    
+
                     accumulator[bin2_id] += 1
 
                 if not accumulator:
@@ -617,13 +619,13 @@ class TabixAggregator(ContactBinner):
                         columns=['bin1_id', 'bin2_id', 'count'])
                           .sort_values('bin2_id')
                 )
-                
+
                 accumulator.clear()
-        
+
         logger.info('Finished {}:{}-{}|*'.format(chrom1, start, end))
 
         return pd.concat(rows, axis=0) if len(rows) else None
-    
+
     def __iter__(self):
         granges = balanced_partition(self.gs, self.n_chunks, self.file_contigs)
         for df in self._map(self.aggregate, granges):
@@ -647,7 +649,7 @@ class PairixAggregator(ContactBinner):
         import dill
         import pickle
         dill.settings['protocol'] = pickle.HIGHEST_PROTOCOL
-        
+
         self._map = map
         self.n_chunks = n_chunks
         self.is_one_based = bool(is_one_based)
@@ -659,7 +661,7 @@ class PairixAggregator(ContactBinner):
         self.file_contigs = set(
             itertools.chain.from_iterable(
                 [b.split('|') for b in f.get_blocknames()]))
-        
+
         if not len(self.file_contigs):
             raise RuntimeError("No reference sequences found.")
         for c1, c2 in itertools.combinations(self.file_contigs, 2):
@@ -678,14 +680,14 @@ class PairixAggregator(ContactBinner):
             if self.n_chunks > old_n:
                 logger.info("Pairs file has {} lines. Increasing max-split to {}.".format(
                     n_lines, self.n_chunks))
-        
+
         # all requested contigs will be placed in the output matrix
         self.gs = GenomeSegmentation(chromsizes, bins)
-        
+
         # find available contigs in the contact list
         self.filepath = filepath
         self.n_records = None
-       
+
         # warn about requested contigs not seen in the contact list
         for chrom in self.gs.contigs:
             if chrom not in self.file_contigs:
@@ -707,7 +709,7 @@ class PairixAggregator(ContactBinner):
         C2 = self.C2
         P1 = self.P1
         P2 = self.P2
-        
+
         logger.info('Binning {}:{}-{}|*'.format(chrom1, start, end))
 
         f = pypairix.open(filepath, 'r')
@@ -718,39 +720,39 @@ class PairixAggregator(ContactBinner):
         accumulator = Counter()
         rows = []
         for bin1_id, bin1 in these_bins.iterrows():
-            
+
             for chrom2, cid2 in six.iteritems(remaining_chroms):
-                
+
                 chrom2_size = chromsizes[chrom2]
 
                 if chrom1 != chrom2 and f.exists2(chrom2, chrom1):  # flipped
-                    iterator = f.query2D(chrom2, 0, chrom2_size, 
+                    iterator = f.query2D(chrom2, 0, chrom2_size,
                                          chrom1, bin1.start, bin1.end)
                     pos2_col = P1
                 else:
-                    iterator = f.query2D(chrom1, bin1.start, bin1.end, 
+                    iterator = f.query2D(chrom1, bin1.start, bin1.end,
                                          chrom2, 0, chrom2_size)
                     pos2_col = P2
 
                 for line in iterator:
-                    
+
                     pos2 = int(line[pos2_col]) - decr
 
                     if binsize is None:
                         lo = chrom_binoffset[cid2]
                         hi = chrom_binoffset[cid2 + 1]
                         bin2_id = lo + np.searchsorted(
-                            start_abspos[lo:hi], 
+                            start_abspos[lo:hi],
                             chrom_abspos[cid2] + pos2,
                             side='right') - 1
                     else:
                         bin2_id = chrom_binoffset[cid2] + (pos2 // binsize)
-                    
+
                     accumulator[bin2_id] += 1
-            
+
             if not accumulator:
                 continue
-            
+
             rows.append(
                 pd.DataFrame({
                     'bin1_id': bin1_id,
@@ -759,13 +761,13 @@ class PairixAggregator(ContactBinner):
                     columns=['bin1_id', 'bin2_id', 'count'])
                       .sort_values('bin2_id')
             )
-            
+
             accumulator.clear()
-        
+
         logger.info('Finished {}:{}-{}|*'.format(chrom1, start, end))
 
         return pd.concat(rows, axis=0) if len(rows) else None
-    
+
     def __iter__(self):
         granges = balanced_partition(self.gs, self.n_chunks, self.file_contigs)
         for df in self._map(self.aggregate, granges):
@@ -790,7 +792,7 @@ class SparseLoader(ContactBinner):
         ('count', int),
     ])
 
-    def __init__(self, filepath, bins, chunksize, field_numbers=None, 
+    def __init__(self, filepath, bins, chunksize, field_numbers=None,
                  field_dtypes=None, one_based=False):
         """
         Parameters
@@ -812,7 +814,7 @@ class SparseLoader(ContactBinner):
             self.field_numbers.update(field_numbers)
         self.columns = list(self.field_numbers.keys())
         self.usecols = list(self.field_numbers.values())
-        
+
         # Assign the column dtypes. Assume additional value fields are float.
         self.out_columns = ['bin1_id', 'bin2_id', 'count']
         self.dtypes = self.FIELD_DTYPES.copy()
@@ -827,10 +829,10 @@ class SparseLoader(ContactBinner):
 
     def __iter__(self):
         n_bins = self.n_bins
-        
+
         iterator = pd.read_csv(
-            self.filepath, 
-            sep='\t', 
+            self.filepath,
+            sep='\t',
             iterator=True,
             comment='#',
             chunksize=self.chunksize,
@@ -848,10 +850,10 @@ class SparseLoader(ContactBinner):
                         "Found bin ID <= 0. Are you sure bin IDs are one-based?")
                 chunk['bin1_id'] -= 1
                 chunk['bin2_id'] -= 1
-            if (np.any(chunk['bin1_id'] >= n_bins) or 
+            if (np.any(chunk['bin1_id'] >= n_bins) or
                 np.any(chunk['bin2_id'] >= n_bins)):
                 raise ValueError(
-                    "Found a bin ID that exceeds the declared number of bins. " 
+                    "Found a bin ID that exceeds the declared number of bins. "
                     "Check whether your bin table is correct.")
             yield {k: v.values for k,v in six.iteritems(chunk)}
 
@@ -867,7 +869,7 @@ class SparseBlockLoader(ContactBinner):
         self.chunksize = chunksize
         n_chroms = len(chromsizes)
         n_bins = len(bins)
-        
+
         chrom_ids = bins['chrom'].cat.codes
         self.offsets = np.zeros(n_chroms + 1, dtype=int)
         curr_val = 0
@@ -875,9 +877,9 @@ class SparseBlockLoader(ContactBinner):
             self.offsets[curr_val:value + 1] = start
             curr_val = value + 1
         self.offsets[curr_val:] = n_bins
-        
+
         self.mapping = mapping
-        
+
     def select_block(self, chrom1, chrom2):
         try:
             block = self.mapping[chrom1, chrom2]
@@ -893,12 +895,12 @@ class SparseBlockLoader(ContactBinner):
     def __iter__(self):
         n_bins = len(self.bins)
         chromosomes = self.chromosomes
-        
+
         for cid1, chrom1 in enumerate(chromosomes):
             offset = self.offsets[cid1]
             chrom1_nbins = self.offsets[cid1 + 1] - offset
             spans = partition(0, chrom1_nbins, self.chunksize)
-            
+
             for lo, hi in spans:
                 chunks = []
                 for chrom2 in chromosomes[cid1:]:
@@ -908,7 +910,7 @@ class SparseBlockLoader(ContactBinner):
                         continue
                     chunks.append(block.tocsr()[lo:hi, :])
                 X = scipy.sparse.hstack(chunks).tocsr().tocoo()
-                
+
                 i, j, v = X.row, X.col, X.data
                 mask = (offset + i) <= (offset + j)
                 triu_i, triu_j, triu_v = i[mask], j[mask], v[mask]
@@ -935,19 +937,19 @@ class ArrayLoader(ContactBinner):
             raise ValueError("Number of bins must equal the dimenion of the matrix")
         self.array = array
         self.chunksize = chunksize
-    
+
     def __iter__(self):
         n_bins = self.array.shape[0]
         spans = partition(0, n_bins, self.chunksize)
-        
+
         # TRIU sparsify the matrix
         for lo, hi in spans:
             X = self.array[lo:hi, :]
             i, j = np.nonzero(X)
-            
+
             mask = (lo + i) <= j
             triu_i, triu_j = i[mask], j[mask]
-            
+
             yield {
                 'bin1_id': lo + triu_i,
                 'bin2_id': triu_j,
@@ -957,7 +959,7 @@ class ArrayLoader(ContactBinner):
 
 class ArrayBlockLoader(ContactBinner):
     """
-    
+
     """
     def __init__(self, chromsizes, bins, mapping, chunksize):
         bins = check_bins(bins, chromsizes)
@@ -966,7 +968,7 @@ class ArrayBlockLoader(ContactBinner):
         self.chunksize = chunksize
         n_chroms = len(chromsizes)
         n_bins = len(bins)
-        
+
         chrom_ids = bins['chrom'].cat.codes
         self.offsets = np.zeros(n_chroms + 1, dtype=int)
         curr_val = 0
@@ -974,9 +976,9 @@ class ArrayBlockLoader(ContactBinner):
             self.offsets[curr_val:value + 1] = start
             curr_val = value + 1
         self.offsets[curr_val:] = n_bins
-        
+
         self.mapping = mapping
-        
+
     def select_block(self, chrom1, chrom2):
         try:
             block = self.mapping[chrom1, chrom2]
@@ -992,12 +994,12 @@ class ArrayBlockLoader(ContactBinner):
     def __iter__(self):
         n_bins = len(self.bins)
         chromosomes = self.chromosomes
-        
+
         for cid1, chrom1 in enumerate(chromosomes):
             offset = self.offsets[cid1]
             chrom1_nbins = self.offsets[cid1 + 1] - offset
             spans = partition(0, chrom1_nbins, self.chunksize)
-            
+
             for lo, hi in spans:
                 chunks = []
                 for chrom2 in chromosomes[cid1:]:
@@ -1007,7 +1009,7 @@ class ArrayBlockLoader(ContactBinner):
                         continue
                     chunks.append(block[lo:hi, :])
                 X = np.concatenate(chunks, axis=1)
-                
+
                 i, j = np.nonzero(X)
                 mask = (offset + i) <= (offset + j)
                 triu_i, triu_j = i[mask], j[mask]
