+++ setup.py
@@ -47,11 +47,11 @@ def get_long_description():
 
 
 install_requires = [
-    'six', 
-    'numpy>=1.9', 
-    'scipy>=0.16', 
+    'six',
+    'numpy>=1.9',
+    'scipy>=0.16',
     'pandas>=0.19',
-    'h5py>=2.5', 
+    'h5py>=2.5',
     'click>=6.6',
     'pysam>0.8',
     'cytoolz',
@@ -59,6 +59,7 @@ install_requires = [
     'biopython',
     'pyfaidx',
     'pypairix',
+    'asciitree',
 ]
 
 
@@ -70,7 +71,7 @@ tests_require = [
 
 extras_require = {
     'docs': [
-        'Sphinx>=1.1', 
+        'Sphinx>=1.1',
         'numpydoc>=0.5'
     ]
 }
