+++ tests/test_cli_aggregate.py
@@ -18,7 +18,8 @@ from nose.tools import with_setup, set_trace
 from click.testing import CliRunner
 
 from cooler.cli.cload import cload, tabix as cload_tabix
-from cooler.cli.aggregate import coarsen, zoomify, multires_aggregate
+from cooler.cli.coarsen import coarsen
+from cooler.cli.zoomify import zoomify, multires_aggregate
 
 
 if sys.version_info[0] == 3 and sys.version_info[1] == 3:
